# Fix a bug in the previous gelu implementation