**Description**: 
Fix a bug in the previous gelu implementation
That was really a silly bug. I forget to set correct output ptr.

**Motivation and Context**
- Why is this change required? What problem does it solve?
- If it fixes an open issue, please link to the issue here.