@pranavsharma 2020-04-25
Is this implementation faster than the previous one which was based on FastGelu? cc @tianleiwu