@tianleiwu 2020-04-25
Here is perf test on my machine (i7-8700 CPU @ 3.20GHz, 12 logical cores, 6 physical cores):
|              |                  |                 |                 |             |             | 
|--------------|------------------|-----------------|-----------------|-------------|-------------| 
| input_length | intra_op_threads | OMP_NUM_THREADS | OMP_WAIT_POLICY | Avg_Latency (master)| Avg_Latency (this PR)| 
| 98304        | 0                |                 |                 | 0.095       | 0.098       | 
| 1572864      | 0                |                 |                 | 1.883       | 2.06        | 
| 98304        | 0                | 12              | ACTIVE          | 0.092       | 0.099       | 
| 1572864      | 0                | 12              | ACTIVE          | 1.956       | 2.648       | 
| 98304        | 0                | 6               | PASSIVE         | 0.09        | 0.088       | 
| 1572864      | 0                | 6               | PASSIVE         | 1.286       | 2.088       | 

Master version is faster.