@snnn 2020-04-25
@tianleiwu: Yes, in my local test, it shows the similar thing, when input length = 1572864, the master version is about 20% faster than mine, but when input length = 10000, the master version is 50% slower.