@yufenglee 2020-04-26
The typical tensor size is Batch*128*768. If batch == 1, it is 

> @tianleiwu: Yes, in my local test, it shows the similar thing, when input length = 1572864, the master version is about 20% faster than mine, but when input length = 10000, the master version is 50% slower.

@snnn, the typical input size is batch*128*768 = batch * 98304.