@tianleiwu 2020-04-26
My previous test is one model with a single Gelu op. I also have results on Bert base model with Gelu using openmp. Note that BiasGelu fusion is disabled so that the optimized model has Gelu operator. There are two set of settings: one set is default (no OMP setting), another with OMP setting. sequence_length=128 has input length 1572864 for Gelu, and sequence_length=16 has input length 98304.

Among 10 settings below, master is better on majority, and this PR wins only 1.

| sequence_length | OMP_NUM_THREADS | OMP_WAIT_POLICY | avg latency (master) | avg latency (this PR) | 
|-----------------|-----------------|-----------------|------------------|-------------------| 
| 8               |                 |                 | 28.17            | 28.25             | 
| 16              |                 |                 | 32.16            | 32.74             | 
| 32              |                 |                 | 42.37            | 42.86             | 
| 64              |                 |                 | 65.39            | 66.45             | 
| 128             |                 |                 | 110.37           | 111.70            | 
|                 |                 |                 |                  |                   | 
| 8               | 12              | PASSIVE         | 31.27            | 31.31             | 
| 16              | 12              | PASSIVE         | 32.59            | 32.87             | 
| 32              | 12              | PASSIVE         | 41.50            | **40.92**             | 
| 64              | 12              | PASSIVE         | 58.23            | 60.32             | 
| 128             | 12              | PASSIVE         | 101.49           | 102.51            |